multi-model now works. Only issue is support for multi-file output
