+++ raven/processes/wps_raven_multi_model.py
@@ -1,6 +1,6 @@
 from raven.processes import RavenProcess
 from pywps import LiteralInput
-from raven.models import get_model, HMETS, GR4JCN, HBVEC, MOHYSE
+from raven.models import HMETS, GR4JCN, HBVEC, MOHYSE, RavenMultiModel
 from . import wpsio as wio
 
 hmets = LiteralInput('hmets', 'Comma separated list of HMETS parameters',
@@ -14,6 +14,7 @@ gr4jcn = LiteralInput('gr4jcn', 'Comma separated list of GR4JCN parameters',
                       min_occurs=0)
 
 
+# This won't work until we support metalink output for multiple files, unless we zip the output.
 class RavenMultiModelProcess(RavenProcess):
     identifier = 'raven-multi-model'
     abstract = 'Multi model simulation'
@@ -26,35 +27,8 @@ class RavenMultiModelProcess(RavenProcess):
 
     inputs = [wio.ts, hmets, gr4jcn, wio.start_date, wio.end_date, wio.duration, wio.run_name,
               wio.name, wio.area, wio.latitude, wio.longitude, wio.elevation]
+    model_cls = RavenMultiModel
 
-    def __init__(self):
-
-        super(RavenProcess, self).__init__(
-            self._handler,
-            identifier=self.identifier,
-            title=self.title,
-            version=self.version,
-            abstract=self.abstract,
-            inputs=self.inputs,
-            outputs=self.outputs,
-            status_supported=True,
-            store_supported=True
-        )
-
-    def _handler(self, request, response):
-
-        # Extract the individual model parameters
-        params = {}
-        for key, val in self.tuple_inputs.items():
-            if key in request.inputs:
-                params[key] = request.inputs.pop(key)
-
-        # Launch the individual models
-        resp = {}
-        for key, val in params.items():
-            req = request.copy()
-            req['params'] = val
-            self.model_cls = get_model(key)
-            resp[key] = RavenProcess._handler(self, req, response)
-
-        # Aggregate the results
+    def model(self, request):
+        models = list(set(request.inputs.keys()).intersection(self.tuple_inputs.keys()))
+        return self.model_cls(models=models, workdir=self.workdir)
