+++ tests/test_wps_raven_multi_model.py
@@ -43,24 +43,7 @@ class TestRavenMultiModelProcess:
                     )
 
         resp = client.get(
-            service='WPS', request='Execute', version='1.0.0', identifier='raven-gr4j-cemaneige',
+            service='WPS', request='Execute', version='1.0.0', identifier='raven-multi-model',
             datainputs=datainputs)
 
         assert_response_success(resp)
-
-        out = get_output(resp.xml)
-        assert 'diagnostics' in out
-        tmp_file, _ = urlretrieve(out['diagnostics'])
-        tmp_content = open(tmp_file).readlines()
-
-        # checking correctness of NSE (full period 1954-2010 would be NSE=0.511214)
-        assert 'DIAG_NASH_SUTCLIFFE' in tmp_content[0]
-        idx_diag = tmp_content[0].split(',').index("DIAG_NASH_SUTCLIFFE")
-        diag = np.float(tmp_content[1].split(',')[idx_diag])
-        np.testing.assert_almost_equal(diag, -0.130318, 4, err_msg='NSE is not matching expected value')
-
-        # checking correctness of RMSE (full period 1954-2010 would be RMSE=32.8827)
-        assert 'DIAG_RMSE' in tmp_content[0]
-        idx_diag = tmp_content[0].split(',').index("DIAG_RMSE")
-        diag = np.float(tmp_content[1].split(',')[idx_diag])
-        np.testing.assert_almost_equal(diag, 38.1697, 4, err_msg='RMSE is not matching expected value')
